[ re #143 ] First Attempt at a safe subsection
This should all be ok. The next step is to make sure that the tests
run by travis enforce the fact that all the modules in Safe/ are
compatible with --safe.

Rather than writing a crude sed script like I've done locally,
I should probably reuse the ideas in GenerateEverything.hs