[ re #235 ] experiment: list, vec literals
Note that I haven't done List.NonEmpty literals because that would
have needed a lot of changes in the module.

Not sure the change is worth it: we sometimes have to use extra
parentheses (whenever we pass the literal as an argument to a function
or even when we use it as a pattern).